Moved pz layer creation function to utils.
